Add react router and re add old app mount new app on builder url with code splitting
