+++ package.json
@@ -37,6 +37,7 @@
     "react-dom": "^17.0.2",
     "react-dropzone": "^11.0.3",
     "react-icons": "^3.10.0",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "^4.0.3",
     "rollup": "^2.61.1",
     "sparql-http-client": "^2.4.0",
