+++ src/App.js
@@ -1,6 +1,4 @@
 import React, { useState, useCallback, useMemo, useEffect, useRef } from 'react'
-import { BsArrowClockwise } from 'react-icons/bs'
-import { useMonaco } from '@monaco-editor/react'
 import {
   getOptionsConfig,
   getDefaultOptionsValues,
@@ -12,86 +10,41 @@ import Section from './components/Section'
 import Footer from './components/Footer'
 import ScreenSizeAlert from './components/ScreenSizeAlert'
 import DataLoader from './components/DataLoader'
+import ChartSelector from './components/ChartSelector'
 import DataMapping from './components/DataMapping'
 import ChartPreviewWithOptions from './components/ChartPreviewWIthOptions'
 import Exporter from './components/Exporter'
 import get from 'lodash/get'
-import deepEqual from 'fast-deep-equal'
+import find from 'lodash/find'
 import usePrevious from './hooks/usePrevious'
 import { serializeProject } from '@rawgraphs/rawgraphs-core'
+import baseCharts from './charts'
+import useSafeCustomCharts from './hooks/useSafeCustomCharts'
 import useDataLoader from './hooks/useDataLoader'
 import isPlainObject from 'lodash/isPlainObject'
 import CookieConsent from 'react-cookie-consent'
-import useChartBuilder from './hooks/useChartBuilder'
-import CodeChartEditor from './components/CodeChartEditor'
-import useUserChartCode from './hooks/useUserChartCode'
+import CustomChartLoader from './components/CustomChartLoader'
+import CustomChartWarnModal from './components/CustomChartWarnModal'
 
-const INITIAL_CODE = {
-  render: `
-import uuid from 'uuid/dist/umd/uuidv4.min.js'
-import _ from 'lodash'
-
-export default function render(node, data, visualOptions, mapping, originalData, styles) {
-  node.innerHTML = '<div>' +
-    '<h1>Is this the real life? Is this just fantasy? Try 2 edit me...</h1>' +
-    data.map(d => '<h2 style="color:' +  visualOptions.color + '">' + d.name + '</h2>').join('') +
-    '<h3>Some stuff from npm:</h3>' +
-    uuid() + '<br />' + _.repeat('$', 20)
-    '</div>'
-}
-  `.trim(),
-  metadata: `
-export default {
-  id: 'awesome',
-  name: 'Super Awesome',
-  categories: [],
-}
-  `.trim(),
-  dimensions: `
-export default [
-  {
-    id: 'name',
-    name: 'Name',
-    validTypes: ['string'],
-    required: true,
-    operation: 'get',
-  },
-]
-  `.trim(),
-  mapData: `
-export default {
-  name: 'get',
-}
-`.trim(),
-  visualOptions: `
-    export default {
-      color: {
-        type: 'color',
-        label: 'Color',
-        default: '#a41c5b',
-      },
-    }
-  `,
-  chart: `
-import render from './render'
-import metadata from './metadata'
-import dimensions from './dimensions'
-import mapData from './mapData'
-import visualOptions from './visualOptions'
-
-export default {
-  type: 'div',
-  metadata,
-  visualOptions,
-  dimensions,
-  mapData,
-  render,
-}
-  `.trim(),
-  index: `export { default as chart } from './chart'`,
-}
+// #TODO: i18n
 
 function App() {
+  const [
+    customCharts,
+    {
+      toConfirmCustomChart,
+      confirmCustomChartLoad,
+      abortCustomChartLoad,
+      uploadCustomCharts,
+      loadCustomChartsFromUrl,
+      loadCustomChartsFromNpm,
+      importCustomChartFromProject,
+      removeCustomChart,
+      exportCustomChart,
+    },
+  ] = useSafeCustomCharts()
+  const charts = useMemo(() => baseCharts.concat(customCharts), [customCharts])
+
   const dataLoader = useDataLoader()
   const {
     userInput,
@@ -111,25 +64,13 @@ function App() {
     hydrateFromSavedProject,
   } = dataLoader
 
-  // NOTE: This is a tempo workaround in the future i sweart
-  // i can init useDataLoader
-  const initRef = useRef(false)
-  useEffect(() => {
-    if (!initRef.current) {
-      initRef.current = true
-      dataLoader.setUserInput(
-        `
-      text
-      Merry christmas
-      Wo ooo
-    `,
-        { type: 'paste' }
-      )
-    }
-  }, [dataLoader])
-
+  /* From here on, we deal with viz state */
+  const [currentChart, setCurrentChart] = useState(charts[0])
   const [mapping, setMapping] = useState({})
-  const [visualOptions, setVisualOptions] = useState({})
+  const [visualOptions, setVisualOptions] = useState(() => {
+    const options = getOptionsConfig(charts[0]?.visualOptions)
+    return getDefaultOptionsValues(options)
+  })
   const [rawViz, setRawViz] = useState(null)
   const [mappingLoading, setMappingLoading] = useState(false)
   const dataMappingRef = useRef(null)
@@ -147,62 +88,36 @@ function App() {
     }
   }, [])
 
-  const lastChartRef = useRef(null)
-  const syncUIWithChart = useCallback(
-    (nextChart) => {
-      const prevChart = lastChartRef.current
-      // NOTE: This implementation is balanced from time spent
-      // to implemente the right diff algo Vs user experience
-      // we can avoid for example avoid to reset mappings if only label changes
-      // but for now we simply deep compare them =)
-      if (
-        !prevChart ||
-        !deepEqual(prevChart.dimensions, nextChart.dimensions)
-      ) {
-        setMapping({})
-        clearLocalMapping()
-      }
-      if (
-        !prevChart ||
-        !deepEqual(prevChart.visualOptions, nextChart.visualOptions)
-      ) {
-        const options = getOptionsConfig(nextChart?.visualOptions)
-        setVisualOptions(getDefaultOptionsValues(options))
-      }
-      setRawViz(null)
-      lastChartRef.current = nextChart
-    },
-    [clearLocalMapping]
-  )
-  const {
-    code: initialCode,
-    writeCode: writeUserCode,
-    resetCode: resetUserCode,
-  } = useUserChartCode(INITIAL_CODE)
-  const [currentChart, buildChart] = useChartBuilder(null, {
-    onBuilded: syncUIWithChart,
+  // NOTE: When we run the import we want to use the "last"
+  // version of importProject callback
+  const lasImportProjectRef = useRef()
+  useEffect(() => {
+    lasImportProjectRef.current = importProject
   })
-
-  const handleCodeChange = useCallback(
-    (code) => {
-      buildChart(code)
-      writeUserCode(code)
-    },
-    [buildChart, writeUserCode]
-  )
-
-  // NOTE: Woraround cause i am a lazy boy
-  const [editorResetKey, setEditorResetKey] = useState(0)
-  const monaco = useMonaco()
-  const resetCode = useCallback(() => {
-    if (monaco) {
-      monaco.editor.getModels().forEach((e) => {
-        e.dispose()
-      })
+  useEffect(() => {
+    const projectUrlStr = new URLSearchParams(window.location.search).get('url')
+    let projectUrl
+    try {
+      projectUrl = new URL(projectUrlStr)
+    } catch (e) {
+      // BAD URL
+      return
     }
-    resetUserCode()
-    setEditorResetKey((k) => k + 1)
-  }, [monaco, resetUserCode])
+    fetch(projectUrl)
+      .then((r) => (r.ok ? r.text() : Promise.reject(r)))
+      .then(
+        (projectStr) => {
+          const project = deserializeProject(projectStr, baseCharts)
+          const lastImportProject = lasImportProjectRef.current
+          if (lastImportProject) {
+            lastImportProject(project, true)
+          }
+        },
+        (err) => {
+          console.log(`Can't load ${projectUrl}`, err)
+        }
+      )
+  }, [])
 
   //resetting mapping when column names changes (ex: separator change in parsing)
   useEffect(() => {
@@ -223,27 +138,39 @@ function App() {
 
   // update current chart when the related custom charts change under the hood
   // if the related custom chart is removed set the first chart
-  // useEffect(() => {
-  //   if (currentChart.rawCustomChart) {
-  //     const currentCustom = find(
-  //       customCharts,
-  //       (c) => c.metadata.id === currentChart.metadata.id
-  //     )
-  //     if (!currentCustom) {
-  //       setCurrentChart(baseCharts[0])
-  //       return
-  //     }
-  //     if (
-  //       currentCustom.rawCustomChart.source !==
-  //       currentChart.rawCustomChart.source
-  //     ) {
-  //       setCurrentChart(currentCustom)
-  //     }
-  //   }
-  // }, [customCharts, currentChart])
+  useEffect(() => {
+    if (currentChart.rawCustomChart) {
+      const currentCustom = find(
+        customCharts,
+        (c) => c.metadata.id === currentChart.metadata.id
+      )
+      if (!currentCustom) {
+        setCurrentChart(baseCharts[0])
+        return
+      }
+      if (
+        currentCustom.rawCustomChart.source !==
+        currentChart.rawCustomChart.source
+      ) {
+        setCurrentChart(currentCustom)
+      }
+    }
+  }, [customCharts, currentChart])
+
+  const handleChartChange = useCallback(
+    (nextChart) => {
+      setMapping({})
+      clearLocalMapping()
+      setCurrentChart(nextChart)
+      const options = getOptionsConfig(nextChart?.visualOptions)
+      setVisualOptions(getDefaultOptionsValues(options))
+      setRawViz(null)
+    },
+    [clearLocalMapping]
+  )
 
   const exportProject = useCallback(async () => {
-    const customChart = null
+    const customChart = await exportCustomChart(currentChart)
     return serializeProject({
       userInput,
       userData,
@@ -280,32 +207,33 @@ function App() {
     visualOptions,
     unstackedColumns,
     unstackedData,
+    exportCustomChart,
   ])
 
   // project import
   const importProject = useCallback(
     async (project, fromUrl) => {
-      let nextCurrentChart = null
-      // if (project.currentChart.rawCustomChart) {
-      //   try {
-      //     nextCurrentChart = await importCustomChartFromProject(
-      //       project.currentChart
-      //     )
-      //   } catch (err) {
-      //     if (err.isAbortByUser) {
-      //       if (fromUrl) {
-      //         // NOTE: clean the url when the user abort loading custom js
-      //         window.history.replaceState(null, null, '/')
-      //       }
-      //       return
-      //     }
-      //     throw err
-      //   }
-      // } else {
-      //   nextCurrentChart = project.currentChart
-      // }
+      let nextCurrentChart
+      if (project.currentChart.rawCustomChart) {
+        try {
+          nextCurrentChart = await importCustomChartFromProject(
+            project.currentChart
+          )
+        } catch (err) {
+          if (err.isAbortByUser) {
+            if (fromUrl) {
+              // NOTE: clean the url when the user abort loading custom js
+              window.history.replaceState(null, null, '/')
+            }
+            return
+          }
+          throw err
+        }
+      } else {
+        nextCurrentChart = project.currentChart
+      }
       hydrateFromSavedProject(project)
-      // setCurrentChart(nextCurrentChart)
+      setCurrentChart(nextCurrentChart)
       setMapping(project.mapping)
       // adding "annotations" for color scale:
       // we annotate the incoming options values (complex ones such as color scales)
@@ -320,26 +248,33 @@ function App() {
       })
       setVisualOptions(project.visualOptions)
     },
-    [hydrateFromSavedProject]
+    [hydrateFromSavedProject, importCustomChartFromProject]
   )
 
   return (
     <div className="App">
       <Header menuItems={HeaderItems} />
+      <CustomChartWarnModal
+        toConfirmCustomChart={toConfirmCustomChart}
+        confirmCustomChartLoad={confirmCustomChartLoad}
+        abortCustomChartLoad={abortCustomChartLoad}
+      />
       <div className="app-sections">
         <Section title={`1. Load your data`} loading={loading}>
           <DataLoader {...dataLoader} hydrateFromProject={importProject} />
         </Section>
         {data && (
-          <Section title="2. Write Your Chart">
-            <button className="btn btn-sm btn-primary mb-2" onClick={resetCode}>
-              <BsArrowClockwise className="mr-2" />
-              Reset
-            </button>
-            <CodeChartEditor
-              key={editorResetKey}
-              initialCode={initialCode}
-              onCodeChange={handleCodeChange}
+          <Section title="2. Choose a chart">
+            <CustomChartLoader
+              loadCustomChartsFromNpm={loadCustomChartsFromNpm}
+              loadCustomChartsFromUrl={loadCustomChartsFromUrl}
+              uploadCustomCharts={uploadCustomCharts}
+            />
+            <ChartSelector
+              onRemoveCustomChart={removeCustomChart}
+              availableCharts={charts}
+              currentChart={currentChart}
+              setCurrentChart={handleChartChange}
             />
           </Section>
         )}
