+++ src/index.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './styles/index.scss'
-import App from './App'
+import WebApp from './WebApp'
 import * as serviceWorker from './serviceWorker'
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <WebApp />
   </React.StrictMode>,
   document.getElementById('root')
 )
